movies/requirements.txt
django
gunicorn
django-heroku
numpy
joblib
djangorestframework
sklearn